Replaced Protean Container Builder with DICBC.
