--- src/Protean/Container/BuilderInterface.php
@@ -1,11 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Neighborhoods\Prefab\Protean\Container;
-
-use Psr\Container\ContainerInterface;
-
-interface BuilderInterface
-{
-    public function build(): ContainerInterface;
-}
