--- src/Symfony/Component/DependencyInjection/ContainerBuilder/FacadeInterface.php
@@ -1,18 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Neighborhoods\Prefab\Symfony\Component\DependencyInjection\ContainerBuilder;
-
-use Symfony\Component\DependencyInjection\ContainerBuilder;
-use Symfony\Component\Finder\Finder;
-
-interface FacadeInterface
-{
-    public function addFinder(Finder $finder): Facade;
-
-    public function getContainerBuilder(): ContainerBuilder;
-
-    public function build(): FacadeInterface;
-
-    public function assembleYaml(): FacadeInterface;
-}
